+++ mirrord/cli/src/execution.rs
@@ -6,7 +6,8 @@ use std::{
 
 use mirrord_analytics::{AnalyticsError, AnalyticsReporter, Reporter};
 use mirrord_config::{
-    config::ConfigError, internal_proxy::MIRRORD_INTPROXY_CONNECT_TCP_ENV, LayerConfig,
+    config::ConfigError, feature::env::mapper::EnvVarsRemapper,
+    internal_proxy::MIRRORD_INTPROXY_CONNECT_TCP_ENV, LayerConfig,
 };
 use mirrord_intproxy::agent_conn::AgentConnectInfo;
 use mirrord_operator::client::OperatorSession;
@@ -164,6 +165,16 @@ impl MirrordExecution {
     /// 1. Drop this struct when an error occurs,
     /// 2. Successfully `exec`,
     /// 3. Wait for intproxy exit with [`MirrordExecution::wait`].
+    ///
+    /// # `async` shenanigans when using the mirrord operator.
+    ///
+    /// Here we connect a websocket to the operator created agent, and this connection
+    /// might get reset if we don't drive its IO (call `await` on the runtime after the
+    /// websocket is up). This is an issue because we start things with `execv`, so we're
+    /// kinda out of the whole Rust world of nicely dropping things.
+    ///
+    /// tl;dr: In `exec_process`, you need to call and `await` either
+    /// [`tokio::time::sleep`] or [`tokio::task::yield_now`] after calling this function.
     #[tracing::instrument(level = Level::TRACE, skip_all)]
     pub(crate) async fn start<P>(
         config: &LayerConfig,
@@ -492,6 +503,12 @@ impl MirrordExecution {
                 .map_err(|error| CliError::EnvFileAccessError(file.clone(), error))?
                 .store;
             env_vars.extend(envs_from_file);
+        }
+
+        if let Some(mapping) = config.feature.env.mapping.clone() {
+            env_vars = EnvVarsRemapper::new(mapping, env_vars)
+                .expect("Failed creating regex, this should've been caught when verifying config!")
+                .remapped();
         }
 
         if let Some(overrides) = &config.feature.env.r#override {